Correctly implement the delivery field in the logs
Using logrus Hook interface. Commit fdefa68 was insufficient.
